Completed drawing timeline for each squib group
A unique timeline is drawn for each squib group the user has created,
and each timeline contains the overlay for the time period during which
that group is firing.  Still a work in progress, with quite  a few
issues to shore up.  The idea is that eventually the user will be able
to drag individual play overlays to wherever they want on the timeline.