Created object to store Squib group data
Created a new object that stores the user defined squib group data,
including the squib group's universe and it's list of times during which
it plays.  The Sequence object now has a list of these SquibGroups, and
the first element in the list should always be the universe squibgroup.
There were lots of changes here, hopefully it doesn't cause any
conflicts.