Added timelines to SquibGroup object
As we briefly discussed, I added individual timelines to each SquibGroup
object.  The idea behind this is to *in the future* allow two groups to
run two different animations simultaneously. With our previous
implementation of a single master timeline, I didn't see a way to do
this (without rewriting all the sequence generation functions).  For now
I created an insertBlanks function in Sequence inorder to pad all groups
not currently getting an animation applied to it, in order to keep all
groups with the same length sequence.  Also added a mergeTimelines to
combine all the SquibGroup timelines into the master timeline.  It
currently doesn't do any checking and can therefore violate the physical
timing restrictions on the system.