Partial Mouse Panning
The scrollpane wasn't working so i kept working with what i had instead.
On the screen where you can view the squibs firing I have the container
moveable with what I've currently got. Just figuring out how to
delete/cover up previous image. Once that is working I'll change the
navigation to update as you move the mouse and not just when the mouse
is released. Then ill add the panning to all the applicable screens and
probably add a reset button that positions it back to the origin.