Universe Visual Update
Click on a squib to select it. All currently selected squibs can be
moved around. For now, click on the red rectangle to deselect all
squibs. When no squibs are selected, everything is moved together.
Still need to change how the squibs move and will add some more
functionality to it.