Setup error feed back from Hardware
Set error feedback from hardware to be printed in GUI instead of only
having output in the console.

This ended up being a PITA, because to get the updates as they occur,
the running of the sequence had to be moved to it's own thread, and the
FX gui set to a listen to the thread.

On the brighter side, this means the program no longer freezes when
clicking send to universe.

Also, added a rough way to check packets received back from universe.
If they don't start with AA (the header), the program returns from the
current timestep with an error.