Updated universe timing
Changed the serial communication class to wait for pong packet before
sending the next round of ping packets.  Before we were sending extra
ping packets, where the firebox would already be charged but we hadn't
processed the response yet so we sent another ping.  Now we stop pinging
as soon as we receive a charged ping.